Refactor game into multiple modules
