+++ constants.py
@@ -1,0 +1,10 @@
+SCREEN_WIDTH = 1024
+SCREEN_HEIGHT = 768
+SCREEN_TITLE = "Grids Prototype"
+
+ROWS = 7
+COLUMNS = 10
+CELL_SIZE = 64
+GRID_WIDTH = COLUMNS * CELL_SIZE
+GRID_HEIGHT = ROWS * CELL_SIZE
+UI_PANEL_WIDTH = SCREEN_WIDTH - GRID_WIDTH
