+++ entities.py
@@ -1,0 +1,7 @@
+class GameEntity:
+    def __init__(self, row, col):
+        self.row = row
+        self.col = col
+
+    def draw(self):
+        pass
